+++ src/Microsoft.VisualStudio.Editor.Razor/DefaultTagHelperFactsService.cs
@@ -4,8 +4,8 @@
 using System;
 using System.Collections.Generic;
 using System.ComponentModel.Composition;
+using System.Linq;
 using Microsoft.AspNetCore.Razor.Language;
-using Microsoft.CodeAnalysis.Razor;
 
 namespace Microsoft.VisualStudio.Editor.Razor
 {
@@ -13,15 +13,6 @@ namespace Microsoft.VisualStudio.Editor.Razor
     [Export(typeof(TagHelperFactsService))]
     internal class DefaultTagHelperFactsService : TagHelperFactsService
     {
-        private readonly TagHelperFactsServiceInternal _tagHelperFactsService;
-
-        [ImportingConstructor]
-        public DefaultTagHelperFactsService(VisualStudioWorkspaceAccessor workspaceAccessor)
-        {
-            var razorLanguageServices = workspaceAccessor.Workspace.Services.GetLanguageServices(RazorLanguage.Name);
-            _tagHelperFactsService = razorLanguageServices.GetRequiredService<TagHelperFactsServiceInternal>();
-        }
-
         public override TagHelperBinding GetTagHelperBinding(
             TagHelperDocumentContext documentContext,
             string tagName,
@@ -29,7 +20,32 @@ namespace Microsoft.VisualStudio.Editor.Razor
             string parentTag,
             bool parentIsTagHelper)
         {
-            return _tagHelperFactsService.GetTagHelperBinding(documentContext, tagName, attributes, parentTag, parentIsTagHelper);
+            if (documentContext == null)
+            {
+                throw new ArgumentNullException(nameof(documentContext));
+            }
+
+            if (tagName == null)
+            {
+                throw new ArgumentNullException(nameof(tagName));
+            }
+
+            if (attributes == null)
+            {
+                throw new ArgumentNullException(nameof(attributes));
+            }
+
+            var descriptors = documentContext.TagHelpers;
+            if (descriptors == null || descriptors.Count == 0)
+            {
+                return null;
+            }
+
+            var prefix = documentContext.Prefix;
+            var tagHelperBinder = new TagHelperBinder(prefix, descriptors);
+            var binding = tagHelperBinder.GetBinding(tagName, attributes.ToList(), parentTag, parentIsTagHelper);
+
+            return binding;
         }
 
         public override IEnumerable<BoundAttributeDescriptor> GetBoundTagHelperAttributes(
@@ -37,7 +53,37 @@ namespace Microsoft.VisualStudio.Editor.Razor
             string attributeName,
             TagHelperBinding binding)
         {
-            return _tagHelperFactsService.GetBoundTagHelperAttributes(documentContext, attributeName, binding);
+            if (documentContext == null)
+            {
+                throw new ArgumentNullException(nameof(documentContext));
+            }
+
+            if (attributeName == null)
+            {
+                throw new ArgumentNullException(nameof(attributeName));
+            }
+
+            if (binding == null)
+            {
+                throw new ArgumentNullException(nameof(binding));
+            }
+
+            var matchingBoundAttributes = new List<BoundAttributeDescriptor>();
+            foreach (var descriptor in binding.Descriptors)
+            {
+                foreach (var boundAttributeDescriptor in descriptor.BoundAttributes)
+                {
+                    if (TagHelperMatchingConventions.CanSatisfyBoundAttribute(attributeName, boundAttributeDescriptor))
+                    {
+                        matchingBoundAttributes.Add(boundAttributeDescriptor);
+
+                        // Only one bound attribute can match an attribute
+                        break;
+                    }
+                }
+            }
+
+            return matchingBoundAttributes;
         }
 
         public override IReadOnlyList<TagHelperDescriptor> GetTagHelpersGivenTag(
@@ -45,12 +91,76 @@ namespace Microsoft.VisualStudio.Editor.Razor
             string tagName,
             string parentTag)
         {
-            return _tagHelperFactsService.GetTagHelpersGivenTag(documentContext, tagName, parentTag);
+            if (documentContext == null)
+            {
+                throw new ArgumentNullException(nameof(documentContext));
+            }
+
+            if (tagName == null)
+            {
+                throw new ArgumentNullException(nameof(tagName));
+            }
+
+            var matchingDescriptors = new List<TagHelperDescriptor>();
+            var descriptors = documentContext?.TagHelpers;
+            if (descriptors?.Count == 0)
+            {
+                return matchingDescriptors;
+            }
+
+            var prefix = documentContext.Prefix ?? string.Empty;
+            if (!tagName.StartsWith(prefix, StringComparison.OrdinalIgnoreCase))
+            {
+                // Can't possibly match TagHelpers, it doesn't start with the TagHelperPrefix.
+                return matchingDescriptors;
+            }
+
+            var tagNameWithoutPrefix = tagName.Substring(prefix.Length);
+            for (var i = 0; i < descriptors.Count; i++)
+            {
+                var descriptor = descriptors[i];
+                foreach (var rule in descriptor.TagMatchingRules)
+                {
+                    if (TagHelperMatchingConventions.SatisfiesTagName(tagNameWithoutPrefix, rule) &&
+                        TagHelperMatchingConventions.SatisfiesParentTag(parentTag, rule))
+                    {
+                        matchingDescriptors.Add(descriptor);
+                        break;
+                    }
+                }
+            }
+
+            return matchingDescriptors;
         }
 
         public override IReadOnlyList<TagHelperDescriptor> GetTagHelpersGivenParent(TagHelperDocumentContext documentContext, string parentTag)
         {
-            return _tagHelperFactsService.GetTagHelpersGivenParent(documentContext, parentTag);
+            if (documentContext == null)
+            {
+                throw new ArgumentNullException(nameof(documentContext));
+            }
+
+            var matchingDescriptors = new List<TagHelperDescriptor>();
+            var descriptors = documentContext?.TagHelpers;
+            if (descriptors?.Count == 0)
+            {
+                return matchingDescriptors;
+            }
+
+            for (var i = 0; i < descriptors.Count; i++)
+            {
+                var descriptor = descriptors[i];
+                foreach (var rule in descriptor.TagMatchingRules)
+                {
+                    if (TagHelperMatchingConventions.SatisfiesParentTag(parentTag, rule))
+                    {
+                        matchingDescriptors.Add(descriptor);
+                        break;
+                    }
+                }
+            }
+
+            return matchingDescriptors;
         }
     }
 }
